+++ bundles/org.eclipse.swt/Eclipse SWT Accessibility/cocoa/org/eclipse/swt/accessibility/SWTAccessibleDelegate.java
@@ -27,10 +27,10 @@ class SWTAccessibleDelegate extends NSObject {
 	static final byte[] SWT_OBJECT = {'S', 'W', 'T', '_', 'O', 'B', 'J', 'E', 'C', 'T', '\0'};
 
 	static Callback accessible2Args, accessible3Args, accessible4Args;
-	static int /*long*/ proc2Args, proc3Args, proc4Args;
+	static long /*int*/ proc2Args, proc3Args, proc4Args;
 
 	Accessible accessible;
-	int /*long*/ delegateJniRef;
+	long /*int*/ delegateJniRef;
 	int childID;
 
 	NSArray attributeNames = null;
@@ -63,7 +63,7 @@ class SWTAccessibleDelegate extends NSObject {
 		byte[] types = {'*','\0'};
 		int size = C.PTR_SIZEOF, align = C.PTR_SIZEOF == 4 ? 2 : 3;
 
-		int /*long*/ cls = OS.objc_allocateClassPair(OS.class_NSObject, className, 0);
+		long /*int*/ cls = OS.objc_allocateClassPair(OS.class_NSObject, className, 0);
 		OS.class_addIvar(cls, SWT_OBJECT, size, (byte)align, types);
 
 		// Add the NSAccessibility overrides
@@ -172,7 +172,7 @@ class SWTAccessibleDelegate extends NSObject {
 		return new NSArray(temp2.id);
 	}
 	
-	static int /*long*/ accessibleProc(int /*long*/ id, int /*long*/ sel) {
+	static long /*int*/ accessibleProc(long /*int*/ id, long /*int*/ sel) {
 		SWTAccessibleDelegate swtAcc = getAccessibleDelegate(id);
 		if (swtAcc == null) return 0;
 		
@@ -196,7 +196,7 @@ class SWTAccessibleDelegate extends NSObject {
 		return 0;
 	}
 
-	static int /*long*/ accessibleProc(int /*long*/ id, int /*long*/ sel, int /*long*/ arg0) {
+	static long /*int*/ accessibleProc(long /*int*/ id, long /*int*/ sel, long /*int*/ arg0) {
 		SWTAccessibleDelegate swtAcc = getAccessibleDelegate(id);
 		if (swtAcc == null) return 0;
 		
@@ -224,7 +224,7 @@ class SWTAccessibleDelegate extends NSObject {
 		return 0;
 	}
 
-	static int /*long*/ accessibleProc(int /*long*/ id, int /*long*/ sel, int /*long*/ arg0, int /*long*/ arg1) {
+	static long /*int*/ accessibleProc(long /*int*/ id, long /*int*/ sel, long /*int*/ arg0, long /*int*/ arg1) {
 		SWTAccessibleDelegate swtAcc = getAccessibleDelegate(id);
 		if (swtAcc == null) return 0;
 		
@@ -242,9 +242,9 @@ class SWTAccessibleDelegate extends NSObject {
 		return 0;
 	}
 
-	static SWTAccessibleDelegate getAccessibleDelegate(int /*long*/ id) {
+	static SWTAccessibleDelegate getAccessibleDelegate(long /*int*/ id) {
 		if (id == 0) return null;
-		int /*long*/ [] jniRef = new int /*long*/ [1];
+		long /*int*/ [] jniRef = new long /*int*/ [1];
 		OS.object_getInstanceVariable(id, SWT_OBJECT, jniRef);
 		if (jniRef[0] == 0) return null;
 		return (SWTAccessibleDelegate)OS.JNIGetObject(jniRef[0]);
